backend-django/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0

# State Machine
django-fsm==2.8.1

# File Storage
django-storages==1.14.2

# PDF Generation
reportlab==4.0.7

# Environment Variables
python-dotenv==1.0.0

# AI Integration
openai==1.12.0

# Utilities
Pillow==10.2.0

